Load dataset from Airbnb New York City dataset to table AB_NYC_6 in postgresql
The changes are to download Airbnb New York City dataset in .csv format to the local machine and ingest data from this dataset to postgresql table AB_NYC_6. The table is created with the appropriate schema to store the dataset and dataset is fetched from csv file to the table.